// LDC.Atlas.WebApp/Client/app/shared/entities/child-sections-costs-to-adjust.entity.ts
export class ChildSectionsCostsToAdjust {
    costId: number;
    sectionId: number;
    costTypeCode: string;
    contractReference: string;
    quantityCode: string;
    quantity: number;
    costDirectionId: number;
    rate: number;
    invoicePercent: number;
    invoicingStatusId: number;
    costDirection: string;
    costIds: number[];
}